requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
boto3==1.34.64
botocore==1.34.64
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.4.4
Django==5.0
django-celery-results==2.5.1
django-environ==0.11.2
django-redis==5.4.0
django-rest-passwordreset==1.4.0
django-storages==1.14.2
django_db==0.0.7
djangorestframework==3.14.0
iniconfig==2.0.0
jmespath==1.0.1
kombu==5.3.5
nose==1.3.7
packaging==24.0
pillow==10.2.0
pluggy==1.4.0
prompt-toolkit==3.0.43
psycopg==3.1.16
psycopg-binary==3.1.16
psycopg-pool==3.2.0
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
pytz==2023.3.post1
redis==5.0.2
s3transfer==0.10.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
